Removed closing PHP tages on PHP-ONLY files
This is bad because it can break PHP in certain situations where there is whitespace and nothing else (empty lines) after the closing tag -- e.g. if you send a REDIRECT header to the user, they will be redirected in php. PHP will then continue to parse the file, and when it gets to the whitespace after a closing tag, it will try to send that. this breaks PHP because the client has already fucked off